# statcache: faster LFU