--- a/statistics/handle/cache/internal/lfu/BUILD.bazel
+++ b/statistics/handle/cache/internal/lfu/BUILD.bazel
@@ -1,0 +1,33 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "lfu",
+    srcs = [
+        "key_set.go",
+        "key_set_shard.go",
+        "lfu_cache.go",
+    ],
+    importpath = "github.com/pingcap/tidb/statistics/handle/cache/internal/lfu",
+    visibility = ["//statistics/handle/cache:__subpackages__"],
+    deps = [
+        "//statistics",
+        "//statistics/handle/cache/internal",
+        "//util/intest",
+        "@com_github_dgraph_io_ristretto//:ristretto",
+        "@org_golang_x_exp//maps",
+    ],
+)
+
+go_test(
+    name = "lfu_test",
+    timeout = "short",
+    srcs = ["lfu_cache_test.go"],
+    embed = [":lfu"],
+    flaky = True,
+    race = "on",
+    shard_count = 6,
+    deps = [
+        "//statistics/handle/cache/internal/testutil",
+        "@com_github_stretchr_testify//require",
+    ],
+)