<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #45367

Problem Summary:

### What is changed and how it works?

```
BenchmarkCachePut
BenchmarkCachePut/LRU
BenchmarkCachePut/LRU-8         	   35224	     65941 ns/op
BenchmarkCachePut/LFU
BenchmarkCachePut/LFU-8         	  143685	     10771 ns/op

BenchmarkCachePutGet/LRU
BenchmarkCachePutGet/LRU-8         	   30358	     40205 ns/op
BenchmarkCachePutGet/LFU
BenchmarkCachePutGet/LFU-8         	   85431	     14172 ns/op
```
### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

add benchmark

- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```